+++ final_pjt_front/src/components/FreeCommentList.vue
@@ -52,6 +52,12 @@ export default {
       this.getFreeComments()
     },
     editComment(comment, updatedCommentData) {
+      if (comment.user !== this.$store.state.userInfo?.pk) {
+        // 작성자가 아닌 경우 알림을 띄웁니다.
+        alert("작성자만 수정할 수 있습니다.");
+        return;
+      }
+
       axios({
         method: 'put',
         url: `${API_URL}/community/free/${this.$route.params.freepk}/${comment.id}/`,
@@ -75,6 +81,12 @@ export default {
         })
     },
     deleteComment(comment) {
+      if (comment.user !== this.$store.state.userInfo?.pk) {
+        // 작성자가 아닌 경우 알림을 띄웁니다.
+        alert("작성자만 삭제할 수 있습니다.");
+        return;
+      }
+
       axios({
         method: 'delete',
         url: `${API_URL}/community/free/${this.$route.params.freepk}/${comment.id}/`,
